Реализована ViewModel, пока в общих чертах
