+++ main_screen/src/main/java/chargeit/main_screen/di/modules.kt
@@ -1,7 +1,7 @@
 package chargeit.main_screen.di
 
 import chargeit.main_screen.settings.*
-import chargeit.main_screen.view.MapsFragmentViewModel
+import chargeit.main_screen.ui.MapsFragmentViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.core.annotation.KoinReflectAPI
 import org.koin.dsl.module
