+++ main_screen/src/main/java/chargeit/main_screen/domain/MapsFragmentDataset.kt
@@ -1,1 +1,7 @@
 package chargeit.main_screen.domain
+
+import com.google.android.gms.maps.model.LatLng
+
+data class MapsFragmentDataset(
+    val examplePlace: LatLng = LatLng(57.756310, 40.969853)
+)
