+++ main_screen/src/main/java/chargeit/main_screen/ui/MapsFragment.kt
@@ -1,57 +1,63 @@
 package chargeit.main_screen.ui
 
-import android.location.Address
+import android.Manifest
+import android.content.Context
+import android.content.pm.PackageManager
+import android.location.*
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.SearchView
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AlertDialog
-import chargeit.core.utils.EMPTY
+import androidx.core.content.ContextCompat
+import androidx.fragment.app.FragmentActivity
+import chargeit.core.utils.ZERO
 import chargeit.core.view.CoreFragment
 import chargeit.main_screen.R
 import chargeit.main_screen.databinding.FragmentMapsBinding
-import chargeit.main_screen.domain.charge_stations.ChargeStation
-import chargeit.main_screen.domain.charge_stations.ChargeStationsState
-import chargeit.main_screen.domain.device_location.DeviceLocation
-import chargeit.main_screen.domain.device_location.DeviceLocationError
-import chargeit.main_screen.domain.device_location.DeviceLocationEvent
-import chargeit.main_screen.domain.device_location.DeviceLocationState
-import chargeit.main_screen.domain.search_addresses.SearchAddress
-import chargeit.main_screen.domain.search_addresses.SearchAddressError
-import chargeit.main_screen.domain.search_addresses.SearchAddressState
+import chargeit.main_screen.domain.MapsFragmentDataset
 import chargeit.main_screen.settings.*
-import chargeit.main_screen.utils.isAtLeastOneGranted
+import chargeit.main_screen.utils.hideKeyboard
+import chargeit.main_screen.utils.makeSnackbar
 import com.google.android.gms.maps.CameraUpdateFactory
 import com.google.android.gms.maps.GoogleMap
 import com.google.android.gms.maps.OnMapReadyCallback
 import com.google.android.gms.maps.SupportMapFragment
+import com.google.android.gms.maps.model.BitmapDescriptorFactory
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.Marker
 import com.google.android.gms.maps.model.MarkerOptions
 import com.google.android.material.bottomsheet.BottomSheetBehavior
+import io.reactivex.rxjava3.disposables.CompositeDisposable
+import kotlinx.android.synthetic.main.fragment_maps.*
+import kotlinx.android.synthetic.main.fragment_maps.view.*
+import kotlinx.coroutines.*
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class MapsFragment : CoreFragment(R.layout.fragment_maps), OnMapReadyCallback {
+class MapsFragment : CoreFragment(R.layout.fragment_maps) {
 
     private var _binding: FragmentMapsBinding? = null
     private val binding get() = _binding!!
     private lateinit var map: GoogleMap
-    private val mapsFragmentViewModel: MapsFragmentViewModel by viewModel()
+    private val mainScreenViewModel: MapsFragmentViewModel by viewModel()
     private var currentAddressMarker: Marker? = null
-    private var currentDeviceLocationMarker: Marker? = null
-    private var locationShowRequestFlag = false
-    private var googlePlayServicesNotPresentErrorFlag = false
-    private var locationIsNotAvailableErrorFlag = false
-    private val defaultLocation = LatLng(DEFAULT_PLACE_LATITUDE, DEFAULT_PLACE_LONGITUDE)
-    private var currentDeviceLocation = defaultLocation
+    private var lastUserLocation: LatLng? = null
+    private var userLocationJob: Job? = null
+    private var addressLocationJob: Job? = null
+    private val viewModelDisposable = CompositeDisposable()
 
-    private val permissionRequest =
-        registerForActivityResult(ActivityResultContracts.RequestMultiplePermissions())
-        { status ->
-            checkPermissionRequestResult(status)
-        }
+    private val coroutineExceptionHandler = CoroutineExceptionHandler { _, throwable ->
+        throwable.printStackTrace()
+    }
+
+    private val scope = CoroutineScope(Dispatchers.IO + coroutineExceptionHandler)
+
+    private val mapReadyCallback = OnMapReadyCallback { googleMap ->
+        map = googleMap
+        initMap()
+        initViewModel()
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
@@ -62,29 +68,143 @@ class MapsFragment : CoreFragment(R.layout.fragment_maps), OnMapReadyCallback {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        binding.map.getFragment<SupportMapFragment>().getMapAsync(this)
-        startAccessToLocation()
+        binding.map.getFragment<SupportMapFragment>().getMapAsync(mapReadyCallback)
         initSearch()
         initButtons()
     }
 
-    override fun onMapReady(googleMap: GoogleMap) {
-        map = googleMap
-        initViewModel()
-        initMap()
-        mapsFragmentViewModel.requestChargeStations()
+    private fun checkPermission(context: FragmentActivity) {
+        when {
+            isPermissionGranted(context, AFL) -> getLocation(context)
+            shouldShowRequestPermissionRationale(AFL) -> showRationaleDialog(context)
+            else -> requestPermission(AFL, REQUEST_CODE)
+        }
     }
 
-    private fun initViewModel() {
-        with(mapsFragmentViewModel) {
-            deviceLocationStateLD.observe(viewLifecycleOwner) { deviceLocationState ->
-                handleDeviceLocationState(deviceLocationState)
+    private fun isPermissionGranted(context: Context, permission: String) =
+        ContextCompat.checkSelfPermission(context, permission) == PackageManager.PERMISSION_GRANTED
+
+    private fun showRationaleDialog(context: Context) {
+        AlertDialog.Builder(context).setTitle(getString(R.string.dialog_rationale_title))
+            .setMessage(getString(R.string.dialog_rationale_message))
+            .setPositiveButton(getString(R.string.dialog_rationale_give_access))
+            { _, _ -> requestPermission(AFL, REQUEST_CODE) }
+            .setNegativeButton(getString(R.string.dialog_rationale_decline))
+            { dialog, _ -> dialog.dismiss() }
+            .create()
+            .show()
+    }
+
+    private fun requestPermission(permission: String, requestCode: Int) {
+        requestPermissions(arrayOf(permission), requestCode)
+    }
+
+    @Deprecated("Deprecated in Java")
+    override fun onRequestPermissionsResult(
+        requestCode: Int, permissions: Array<String>, grantResults: IntArray
+    ) {
+        activity?.let { thisActivity ->
+            checkPermissionsResult(thisActivity, requestCode, grantResults)
+        }
+    }
+
+    private fun checkPermissionsResult(
+        context: FragmentActivity,
+        requestCode: Int,
+        grantResults: IntArray
+    ) {
+        if (requestCode == REQUEST_CODE) {
+            if (grantResults.isNotEmpty()
+                && grantResults.size ==
+                grantResults.count { item -> item == PackageManager.PERMISSION_GRANTED }
+            ) {
+                getLocation(context)
+            } else {
+                showDialog(
+                    context,
+                    title = getString(R.string.dialog_title_no_gps),
+                    message = getString(R.string.dialog_message_no_gps)
+                )
             }
-            searchAddressStateLD.observe(viewLifecycleOwner) { searchAddressState ->
-                handleAddressState(searchAddressState)
+        }
+    }
+
+    private fun showDialog(context: Context, title: String, message: String) {
+        AlertDialog.Builder(context).setTitle(title).setMessage(message)
+            .setNegativeButton(getString(R.string.dialog_button_close))
+            { dialog, _ -> dialog.dismiss() }
+            .create().show()
+    }
+
+    private fun getLocation(context: Context) {
+        if (isPermissionGranted(context, AFL)) {
+            with(map) {
+                isMyLocationEnabled = true
+                uiSettings.isMyLocationButtonEnabled = false
+                uiSettings.isRotateGesturesEnabled = false
             }
-            chargeStationsStateLD.observe(viewLifecycleOwner) { chargeStationsState ->
-                handleChargeStationsState(chargeStationsState)
+            val locationManager =
+                context.getSystemService(Context.LOCATION_SERVICE) as LocationManager
+            if (locationManager.isProviderEnabled(PROVIDER)) {
+                val provider = locationManager.getProvider(PROVIDER)
+                provider?.let {
+                    locationManager.requestLocationUpdates(
+                        PROVIDER,
+                        USER_LOCATION_REFRESH_PERIOD,
+                        USER_LOCATION_MINIMAL_DISTANCE,
+                        object : LocationListener {
+                            override fun onLocationChanged(location: Location) {
+                                lastUserLocation = LatLng(location.latitude, location.longitude)
+                                showCurrentUserLocation(context, location)
+                            }
+
+                            @Deprecated("Deprecated in Java")
+                            override fun onStatusChanged(
+                                provider: String,
+                                status: Int,
+                                extras: Bundle
+                            ) {
+                            }
+
+                            override fun onProviderEnabled(provider: String) {}
+                            override fun onProviderDisabled(provider: String) {}
+                        }
+                    )
+                }
+            } else {
+                val location = locationManager.getLastKnownLocation(PROVIDER)
+                location?.let {
+                    lastUserLocation = LatLng(location.latitude, location.longitude)
+                    showCurrentUserLocation(context, location)
+                    showDialog(
+                        context,
+                        getString(R.string.dialog_title_gps_turned_off),
+                        getString(R.string.dialog_message_last_known_location)
+                    )
+                } ?: run {
+                    showDialog(
+                        context,
+                        getString(R.string.dialog_title_gps_turned_off),
+                        getString(R.string.dialog_message_last_location_unknown)
+                    )
+                }
+            }
+        } else {
+            showRationaleDialog(context)
+        }
+    }
+
+    private fun showCurrentUserLocation(context: Context, location: Location) {
+        userLocationJob?.cancel()
+        userLocationJob = scope.launch {
+            val addresses =
+                Geocoder(context).getFromLocation(
+                    location.latitude,
+                    location.longitude,
+                    USER_LOCATION_SEARCH_RESULTS
+                )
+            withContext(Dispatchers.Main) {
+                showAddress(addresses.component1(), USER_LOCATION_ZOOM_LEVEL, true)
             }
         }
     }
@@ -92,8 +212,17 @@ class MapsFragment : CoreFragment(R.layout.fragment_maps), OnMapReadyCallback {
     private fun initSearch() {
         binding.search.setOnQueryTextListener(object : SearchView.OnQueryTextListener {
             override fun onQueryTextSubmit(query: String?): Boolean {
-                hideKeyboard(requireActivity())
-                mapsFragmentViewModel.checkQuery(query)
+                view?.let { thisView ->
+                    query?.let { thisQuery ->
+                        thisView.hideKeyboard()
+                        val resQuery = thisQuery.lowercase().trim()
+                        if (resQuery.isNotBlank()) {
+                            searchByAddress(resQuery, thisView)
+                        } else {
+                            thisView.makeSnackbar(text = getString(R.string.empty_query_message))
+                        }
+                    }
+                }
                 return true
             }
 
@@ -104,69 +233,26 @@ class MapsFragment : CoreFragment(R.layout.fragment_maps), OnMapReadyCallback {
     private fun initButtons() {
         with(binding) {
             filterScreenButton.setOnClickListener { filterScreenButtonClick() }
-            deviceCoordinatesButton.setOnClickListener { deviceCoordinatesButtonClick() }
+            userCoordinatesButton.setOnClickListener { userCoordinatesButtonClick() }
         }
     }
 
     private fun initMap() {
-        googleMapInterfaceConfig()
-        setMarkersBehavior()
-        showInitialDeviceLocation()
-    }
-
-    private fun shouldShowAtLeastOneRationale() =
-        shouldShowRequestPermissionRationale(COARSE_PERMISSION)
-                || shouldShowRequestPermissionRationale(FINE_PERMISSION)
-
-    private fun checkPermissions() {
-        if (!isAtLeastOneGranted(requireContext())) {
-            if (shouldShowAtLeastOneRationale()) {
-                showRationaleDialog()
-            } else {
-                showNotGrantedNoAskDialog()
-            }
+        userCoordinatesButtonClick()
+        map.setOnMarkerClickListener { marker ->
+            view?.makeSnackbar(text = marker.title ?: getString(R.string.no_title_message))
+            true
         }
     }
 
-    private fun checkErrorFlags() {
-        if (googlePlayServicesNotPresentErrorFlag) {
-            showNoPlayServicesDialog()
-        }
-        if (locationIsNotAvailableErrorFlag) {
-            makeSnackbar(
-                view = binding.root,
-                text = getString(R.string.location_is_not_available_error)
+    private fun initViewModel() {
+        with(mainScreenViewModel) {
+            viewModelDisposable.addAll(
+                datasetLiveData.subscribe { thisDataset ->
+                    handleDataset(thisDataset)
+                }
             )
-        }
-    }
-
-    private fun startAccessToLocation() {
-        if (isAtLeastOneGranted(requireContext())) {
-            startLocationUpdates()
-        } else {
-            permissionRequest.launch(arrayOf(COARSE_PERMISSION, FINE_PERMISSION))
-        }
-    }
-
-    private fun checkPermissionRequestResult(status: Map<String, Boolean>) {
-        val noRationale = !shouldShowAtLeastOneRationale()
-        val isGranted = (status[COARSE_PERMISSION] == true || status[FINE_PERMISSION] == true)
-        when {
-            isGranted -> startLocationUpdates()
-            noRationale -> showNotGrantedNoAskDialog()
-            else -> showRationaleDialog()
-        }
-    }
-
-    private fun startLocationUpdates() {
-        mapsFragmentViewModel.startLocationUpdates()
-        locationShowRequestFlag = true
-    }
-
-    private fun googleMapInterfaceConfig() {
-        with(map) {
-            uiSettings.isMyLocationButtonEnabled = false
-            uiSettings.isRotateGesturesEnabled = false
+            powerOn()
         }
     }
 
@@ -175,197 +261,70 @@ class MapsFragment : CoreFragment(R.layout.fragment_maps), OnMapReadyCallback {
             BottomSheetBehavior.STATE_HALF_EXPANDED
     }
 
-    private fun deviceCoordinatesButtonClick() {
-        checkPermissions()
-        checkErrorFlags()
-        if (currentDeviceLocation != defaultLocation) {
-            showCurrentDeviceLocation()
+    private fun userCoordinatesButtonClick() {
+        lastUserLocation?.let { location ->
+            moveCamera(location, USER_LOCATION_ZOOM_LEVEL, true)
+        } ?: run {
+            val defaultPlace = mainScreenViewModel.getDefaultPlace()
+            moveCamera(defaultPlace.coordinates, defaultPlace.zoomLevel, false)
         }
+        activity?.let { thisActivity -> checkPermission(thisActivity) }
     }
 
-    private fun showInitialDeviceLocation() {
-        moveCamera(currentDeviceLocation, DEFAULT_PLACE_ZOOM_LEVEL, false)
-        locationShowRequestFlag = true
+    private fun showAddress(address: Address, zoomLevel: Float, animated: Boolean) {
+        moveCamera(LatLng(address.latitude, address.longitude), zoomLevel, animated)
     }
 
-    private fun showCurrentDeviceLocation() {
-        moveCamera(currentDeviceLocation, DEVICE_LOCATION_ZOOM_LEVEL, true)
-        locationShowRequestFlag = false
-
-    }
-
-    private fun setMarkersBehavior() {
-        map.setOnMarkerClickListener { marker ->
-            val title = marker.title ?: getString(R.string.no_title_message)
-            makeSnackbar(view = binding.root, text = title)
-            true
-        }
-    }
-
-    private fun moveCamera(location: LatLng, zoomLevel: Float, isAnimated: Boolean) {
+    private fun moveCamera(location: LatLng, zoomLevel: Float, animated: Boolean) {
         val cameraProperties = CameraUpdateFactory.newLatLngZoom(location, zoomLevel)
-        if (isAnimated) {
+        if (animated) {
             map.animateCamera(cameraProperties)
         } else {
             map.moveCamera(cameraProperties)
         }
     }
 
-    private fun showLocationMarker(markerOptions: MarkerOptions) {
-        currentDeviceLocationMarker?.remove()
-        currentDeviceLocationMarker = map.addMarker(markerOptions)
-    }
-
-    private fun showAddressMarker(markerOptions: MarkerOptions) {
-        currentAddressMarker?.remove()
-        currentAddressMarker = map.addMarker(markerOptions)
-    }
-
-    private fun showChargeStationMarker(markerOptions: MarkerOptions) {
-        map.addMarker(markerOptions)
-    }
-
-    private fun showAddress(address: Address, zoomLevel: Float, isAnimated: Boolean) {
-        moveCamera(LatLng(address.latitude, address.longitude), zoomLevel, isAnimated)
-    }
-
-    private fun handleDeviceLocationState(deviceLocationState: DeviceLocationState) {
-        when (deviceLocationState) {
-            is DeviceLocationState.Success -> showLocation(deviceLocationState.deviceLocation)
-            is DeviceLocationState.Error -> processLocationErrors(deviceLocationState.deviceLocationError)
-            is DeviceLocationState.Event -> processLocationEvents(deviceLocationState.deviceLocationEvent)
-            is DeviceLocationState.Loading -> showLocationLoadingStatus()
+    private fun searchByAddress(query: String, view: View) {
+        addressLocationJob?.cancel()
+        addressLocationJob = scope.launch {
+            val addresses =
+                Geocoder(view.context).getFromLocationName(query, MAX_ADDRESS_SEARCH_RESULTS)
+            if (addresses.isNotEmpty()) {
+                val address = addresses.component1()
+                withContext(Dispatchers.Main) {
+                    showAddress(address, ADDRESS_SEARCH_ZOOM_LEVEL, true)
+                    currentAddressMarker?.remove()
+                    currentAddressMarker = map.addMarker(
+                        MarkerOptions().position(LatLng(address.latitude, address.longitude))
+                            .title(address.getAddressLine(Int.ZERO))
+                    )
+                }
+            } else {
+                view.makeSnackbar(text = getString(R.string.address_is_not_found_message))
+            }
         }
     }
 
-    private fun showLocation(deviceLocation: DeviceLocation) {
-        currentDeviceLocation = deviceLocation.markerOptions.position
-        showLocationMarker(deviceLocation.markerOptions)
-        if (locationShowRequestFlag) {
-            showCurrentDeviceLocation()
-        }
-    }
-
-    private fun processLocationErrors(deviceLocationError: DeviceLocationError) {
-        when (deviceLocationError.errorID) {
-            PERMISSION_ERROR_ID -> checkPermissions()
-            GOOGLE_PLAY_SERVICES_NOT_PRESENT_ERROR_ID -> processNoPlayServicesError()
-            LOCATION_IS_NOT_AVAILABLE_ERROR_ID -> showLocationIsNotAvailableDialog()
-        }
-    }
-
-    private fun processNoPlayServicesError() {
-        googlePlayServicesNotPresentErrorFlag = true
-        showNoPlayServicesDialog()
-    }
-
-    private fun processLocationEvents(deviceLocationEvent: DeviceLocationEvent) {
-        if (deviceLocationEvent.eventID == LOCATION_IS_AVAILABLE_EVENT_ID) {
-            locationIsNotAvailableErrorFlag = false
-            makeSnackbar(
-                view = binding.root,
-                text = deviceLocationEvent.message
-            )
-        }
-    }
-
-    private fun showLocationLoadingStatus() {}
-
-    private fun handleAddressState(searchAddressState: SearchAddressState) {
-        when (searchAddressState) {
-            is SearchAddressState.Success -> showFoundAddresses(searchAddressState.searchAddress)
-            is SearchAddressState.Error -> processAddressStateError(searchAddressState.searchAddressError)
-            is SearchAddressState.Loading -> showAddressStateLoadingStatus()
-        }
-    }
-
-    private fun showFoundAddresses(searchAddress: SearchAddress) {
-        val address = searchAddress.address
-        showAddressMarker(searchAddress.markerOptions)
-        showAddress(address, ADDRESS_SEARCH_ZOOM_LEVEL, true)
-    }
-
-    private fun processAddressStateError(searchAddressError: SearchAddressError) {
-        makeSnackbar(
-            view = binding.root,
-            text = searchAddressError.message ?: String.EMPTY
+    private fun handleDataset(dataset: MapsFragmentDataset) {
+        val location = dataset.examplePlace
+        map.addMarker(
+            MarkerOptions()
+                .position(location)
+                .icon(BitmapDescriptorFactory.fromResource(R.drawable.ic_map_marker))
         )
-    }
-
-    private fun showAddressStateLoadingStatus() {}
-
-    private fun handleChargeStationsState(state: ChargeStationsState) {
-        when (state) {
-            is ChargeStationsState.Success -> showChargeStationList(state.chargeStations)
-            is ChargeStationsState.Error -> processChargeStationError(state.chargeStationError)
-            is ChargeStationsState.Loading -> showChargeStationsLoadingStatus()
-        }
-    }
-
-    private fun showChargeStationList(chargeStations: List<ChargeStation>) {
-        chargeStations.forEach { chargeStation ->
-            showChargeStationMarker(chargeStation.markerOptions)
-        }
-    }
-
-    private fun processChargeStationError(error: Throwable) {}
-
-    private fun showChargeStationsLoadingStatus() {}
-
-    private fun showRationaleDialog() {
-        showDialog(
-            title = getString(R.string.rationale_title),
-            message = getString(R.string.rationale_message),
-            positiveButtonText = getString(R.string.dialogs_positive_button)
-        )
-    }
-
-    private fun showNotGrantedNoAskDialog() {
-        showDialog(
-            title = getString(R.string.not_granted_no_ask_title),
-            message = getString(R.string.not_granted_no_ask_message),
-            positiveButtonText = getString(R.string.dialogs_positive_button)
-        )
-    }
-
-    private fun showLocationIsNotAvailableDialog() {
-        locationIsNotAvailableErrorFlag = true
-        showDialog(
-            title = getString(R.string.location_not_available_title),
-            message = getString(R.string.location_not_available_message),
-            positiveButtonText = getString(R.string.dialogs_positive_button)
-        )
-    }
-
-    private fun showNoPlayServicesDialog() {
-        showDialog(
-            title = getString(R.string.no_play_services_title),
-            message = getString(R.string.no_play_services_message),
-            positiveButtonText = getString(R.string.dialogs_positive_button)
-        )
-    }
-
-    private fun showDialog(title: String, message: String, positiveButtonText: String) {
-        AlertDialog.Builder(requireContext())
-            .setTitle(title)
-            .setMessage(message)
-            .setPositiveButton(positiveButtonText)
-            { dialog, _ -> dialog.dismiss() }
-            .create()
-            .show()
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
+        scope.cancel()
+        viewModelDisposable.dispose()
     }
 
     companion object {
-        private const val ADDRESS_SEARCH_ZOOM_LEVEL = 10.0f
-        private const val DEVICE_LOCATION_ZOOM_LEVEL = 15f
-        private const val DEFAULT_PLACE_ZOOM_LEVEL = 10.0f
-        private const val DEFAULT_PLACE_LATITUDE = 55.751513
-        private const val DEFAULT_PLACE_LONGITUDE = 37.616655
+        private const val AFL = Manifest.permission.ACCESS_FINE_LOCATION
+        private const val PROVIDER = LocationManager.NETWORK_PROVIDER
+        private const val REQUEST_CODE = 1
 
         @JvmStatic
         fun newInstance() = MapsFragment()
