--- main_screen/src/main/java/chargeit/main_screen/view/MapsFragmentViewModel.kt
@@ -1,7 +1,0 @@
-package chargeit.main_screen.view
-
-import chargeit.core.viewmodel.CoreViewModel
-
-class MapsFragmentViewModel : CoreViewModel() {
-
-}
